Fix an occasional hitch.
A situation can occur where the main thread believes that there is still a job thread active, so it waits for the work_exhaused condition variable, but there is actually no work active, so the next time that the variable is woken is a full second later by the first thread that fails to queue more work.  Conceptually though, it doesn't make much sense for a thread to wait a full second before checking for more work, so just reduce that time to a millisecond.