Descriptor set copies for vulkan resource table binding.
Previously attempting to update a resource in a table would result in invalidating vulkan command lists because descriptor sets cannot be changed while bound in a pending command list. To fix this, descriptor sets in the table are never bound and instead copied to transient sets used exclusively for that frame.