Shader reflection for opengl.
This makes it possible to get rid of the manual resource meta naming that was necessary for the opengl backend to work. The reflection now aligns reflected resources with the intended binding order of resources.

ShaderConductor is updated to pull in a change that properly assigns bind locations to resources in cross-compiled glsl.

Matrix data has a long standing issue where it may or may not be properly packed. More investigation into this later. For now each float4x4 has to be marked up explicitly with column_major.