Overhaul backbuffer management.
Add a swap chain object and pass the responsibility for creation on to the application.

Framebuffer objects are also bubbled up to the application layer to support creating a framebuffer for each backbuffer image. Patterns surrounding render pass creation are adjusted accordingly.

The latter render passes now draw directly to the backbuffer and the fake present targets are finally deleted.

A wait for idle function is also added for use soon.