Rudimentary hdr support.
Only for the dx12 backend for now, vulkan detection of hdr is a puzzle. Encoding for display is moved to a final render pass, the only one to draw to the backbuffer. As a result, ui now draws in linear space.

imgui is converted to premultiplied alpha in order to blend properly while drawn in linear space.